MODELAGEM_BACKEND: realizando o merge da refatoração do backend devido a alteração na modelagem das entidades.

+++ nutriburger-backend/src/main/java/com/grupo6/nutriburger/repository/ProdutoRepository.java
@@ -7,8 +7,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 
     @Query("SELECT new com.grupo6.nutriburger.dto.ProdutoBasicoDTO(p.id, p.nome, p.preco) FROM Produto p")
     List<ProdutoBasicoDTO> buscarProdutosBasicos();
+
 }
